@@ -1,9 +1,9 @@
 package com.team15app.team15.adapters
 
 import android.os.Bundle
-import android.support.v4.app.Fragment
-import android.support.v4.app.FragmentManager
-import android.support.v4.app.FragmentStatePagerAdapter
+import androidx.fragment.app.Fragment
+import androidx.fragment.app.FragmentManager
+import androidx.fragment.app.FragmentStatePagerAdapter
 import com.team15app.team15.dialogs.MatchInfoFragment
 import com.team15app.team15.dialogs.PlayerJerseyFragment
 
