@@ -1,8 +1,8 @@
 package com.team15app.team15.adapters
 
-import android.support.v4.app.Fragment
-import android.support.v4.app.FragmentManager
-import android.support.v4.app.FragmentStatePagerAdapter
+import androidx.fragment.app.Fragment
+import androidx.fragment.app.FragmentManager
+import androidx.fragment.app.FragmentStatePagerAdapter
 import com.team15app.team15.Player
 import com.team15app.team15.dialogs.PlayerFragment
 import java.util.*
