@@ -1,7 +1,7 @@
 package com.team15app.team15.dialogs
 
 import android.os.Bundle
-import android.support.v4.app.Fragment
+import androidx.fragment.app.Fragment
 import android.text.InputType
 import android.view.LayoutInflater
 import android.view.View
