@@ -4,8 +4,8 @@ import android.app.AlertDialog
 import android.app.Dialog
 import android.content.Context.INPUT_METHOD_SERVICE
 import android.os.Bundle
-import android.support.v4.app.DialogFragment
-import android.support.v4.view.ViewPager
+import androidx.fragment.app.DialogFragment
+import androidx.viewpager.widget.ViewPager
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
