@@ -1,7 +1,7 @@
 package com.team15app.team15.viewpagers
 
 import android.content.Context
-import android.support.v4.view.ViewPager
+import androidx.viewpager.widget.ViewPager
 import android.util.AttributeSet
 
 class PlayerJerseyViewPager : ViewPager {
